[OPENENGSB-769/reactivate_license-check_profile] * added the profile to the pom again

+++ tooling/maven-openengsb-plugin/src/main/java/org/openengsb/tooling/pluginsuite/openengsbplugin/LicenseCheck.java
@@ -144,8 +144,12 @@ public class LicenseCheck extends AbstractOpenengsbMojo {
     }
 
     private void cleanUp() {
-        licenseHeaderFile.delete();
-        tmpPom.delete();
+        if (licenseHeaderFile != null) {
+            licenseHeaderFile.delete();
+        }
+        if (tmpPom != null) {
+            tmpPom.delete();
+        }
     }
 
 }
